requirements.txt
torch==2.1.0
torchvision==0.16.0
segmentation-models-pytorch==0.3.3
gradio==4.44.0
Pillow==10.4.0
numpy==1.26.4
huggingface_hub==0.24.0
